Remove brew support Part I: switch install command
Description
===========

This patch makes the new install process without `brew` the default. The
old implementation is deleted and will no longer be available.

Note
----

The uninstall command still checks and uninstalls `brew` casks. It will
be updated in a second pull request.

Changes
=======

![REMOVE] `brew` dependency in install command
