+++ install/uvm-install/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
 
 [dependencies]
 flexi_logger = "0.9.2"
+indicatif = "0.9.0"
 log = "0.4.5"
 console = "0.6.1"
 serde = "1.0"
