--- install/uvm-install2/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "uvm-install2"
-version = "2.0.1"
-authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
-
-[dependencies]
-flexi_logger = "0.9.2"
-indicatif = "0.9.0"
-log = "0.4.5"
-console = "0.6.1"
-serde = "1.0"
-serde_derive = "1.0"
-uvm_cli = { path = "../../uvm_cli" }
-uvm_core = { path = "../../uvm_core" }
